add(feature): white-label
Surfaces a `WHITE_LABEL` environment variable that can be defined in the `.env` environment file and/or passed via Docker/Docker Compose to:

* Hide various footer links
* Prevent db from seeding with a few user-specific pieces of data

Not intending to draw credit away from you @rajnandan1 by any means, however, as an open-source project, it is common to add ‘white-labeling’ functionality, especially considering the project is licensed under MIT. The default is set to NOT include this environment variable so that attribution, by default, remains. However, it’d be great to get this merged as sometimes due to regulatory/compliance policies, some of these things need disabled from the get-go for specific users.